Add 'openstack' provisioner playbooks & settings
Added OpenStack provisioner playbooks, roles and settings.

It is used for creating and/or removing resources on OpenStack clouds.

Includes:

* provision networks
    - Allows to specify if network is shared
    - Allows to specify timeout for network
* provisiom subnets
    - Allows to specify cidr, dns servers, allocation pool
      and whether DHCP should be enabled
* provision routers
    - Allows to attach network interfaces
    - Allows to specify default gateway
* provision instances
    - Allows to specify config_drive, nics, image and flavor
* provision flavors
    - Allows to specify ram, disk and vcpus
* provision floating IPs
    - Allows to reuse existing floating IPs
    - Allows to specify fixed IP address

Ansible OpenStack modules use shade. Added shade to requirements.

Added custom version of 'os_floating_ip' module that supports
'internal_network' parameter. It should be removed when the
support will be added to the original module.

The style is based on:
* 4 spaces
* module variables are sorted alphabetically
* blank line after each task