Fix os_floating_ip module
'network' parameter is not used at all in the 'os_floating_ip'
module.

This fix it so the creation of the floating ip would be based
on the chosen pool/network.

This also adds 'python-novaclient<3' to the requirements. Using
higher version will cause the os_floating_ip module to fail.