OVB poc implementation for infrared
OVB (openstack virtual baremetal) allows to install OpenStack on the nodes provisioned on the another instance of the openstack.

For now ospd installer fails on PXE boot (overcloud) install due to the know issue (https://bugs.launchpad.net/tripleo/+bug/1507738 ).

The main problem is that ovb cannot be easily integrated into our current playbboks structure. We might need to consider to add separate playbook/installer for that and:
 - provision given number of barematal nodes with openstack provisioner
 - connect baremetal to ONE provision network
 - modify ospd execution and assign specific flavors (maybe).

 * add ovb roles and variables
 * add bmc node
 * modified ospd related playbooks

Change-Id: I1d26fb27c2b50192248b89766aa1774e32857a5e