Do not use USER environment variable in the hosts
Historically the inventory variables was suffixed with
some random string, but using the USER environment variable does not
makes any sense.

We can completely remove the logic, by not maintaining a symbolic link
at all, or save the various older states of the file for debugging reasons.

This patch adds various suffixes to the hosts file based on which part of
the play created it, and manages the hosts symbolic link. So it is the
2th option.

Change-Id: Ic94a0e3d8e2248ac10ea697b7b0f23695f061182