Adding a tox.ini with hacking
hacking is project used by openstack for generic python code check,
it based on flake8 which is the combination of pyflakes and pep8.

pyflakes some case can indicate real coding issues, it not just a
pushing for readable code. It does some semantical analysis as well.

hacking also contains some coding rules, which mostly not just openstack
specific rules.

For the time being we skip all checks where infrared is not confirming,
but in the future most of the rules is planned to be followed.

You can invoke the tox entry by:
$ tox -epep

The library part is excluded, because those modules can be
modules copied from repos not following the same ruleset and eventually
the will be synced again.

Change-Id: I7a53bed70822e97a112b862f8b906934e4101e05