Install python-virtualenv on the undercloud
With ospd-10 nothing pulls the packages required for
building a virtual environment.

We are using a virtual environment on the undercloud node
to learn the overcloud nodes.

The os_server_facts module depends on shade,
which is expected to be installed by the pip task.

The proper venv building also requires gcc and python-devel.

Install happens just after we establish the repo locations.

Change-Id: I749922b1491646576eed8b96aefd573e83458556