'Set selinux state' play added 'gather_facts: no'
When deploy in a pipeline using tags, after the introspection phase
it is no longer possible to access nodes expect undercloud node.
When play from a playbook doesn't have 'gather_facts'
set to 'no' the play implicitly runs 'setup' module even if it is
not under the tag that is being executed. 'setup' module tries to access
all nodes specified in 'hosts' and fails on the nodes that are no accessible.
And consequently jenkins job that runs ansible with the tag fails.
This change fixes ansible run for an 'overcloud' and 'inventory_update' tags
in install.yml playbook

Change-Id: I850e7b88155656d1c665a95cfc4356d857425909