Creation and configuration of heat_stack_owner
According OSPD documentation, in order to run tempest, it requires that the role
heat_stack_owner be created, and set in the tempest.conf.
This will fix all the orchestration tests failing right now.
This is not a 'workaround', since it's documented that it requires this role in
order to run tempest.

Change-Id: Iebdfb2e0cbcfe025924f27741f913adea38c505e