Fix baremetal setups using cidr instead of proper IP
Currently, CIDR (ExternalNetCidr) is used as $IP address
for vlan setup (ip command), which is wrong. We must use
ExternalInterfaceDefaultRoute variable as this
value contains proper IP for UC node, which also
works as external gateway for OC nodes.

Currently, we set IP of vlan interface as IP
of whole subnet (ending with .0/YY).

* also configure static route using cidr as last step

Change-Id: I5378b4c9c7dce6648f49437e813dbf479ec5dfed