Fix default region for OSPD
For some reason, the default region in OSPD is regionOne instead of
the 'used everywhere' RegionOne. Because of this, and tempest default
region is RegionOne, some tests fails, like all the identity v3 tests.
This patch set the identity.region to regionOne.
This won't be necessary once https://review.openstack.org/#/c/336442/
be merged.

Change-Id: Ie61b181e2fc1634541156d72124172e9e7e69c2d