Added 'lvm' playbook for a storage role.
When running Ansible with tags and empty storage backend, ansible
tries to evaluate included playbook file in 'include' statement
in storage role 'main' playbook anyway. It doesn't depend if storage backend
is defined or not. When storage backend is not defined Ansible run
fails with error that it can't load the file 'playbooks/installer/ospd/.yml'.
It happens only when run ansible with tags.
This patch adds a 'noop' playbook that is being used when no other storage
is defined. Ansible just evaluates it and does nothing.

Change-Id: Idb4d1da1ecd44d79a3f48c6493253311f0b3f93a