Increase tail lenght to 60
Currently, tail is not long enought to catch majority of errors
(from my experience). We should increase it to 60 to catch
as much errors as possible right away during CI run. This is not
so costly as there is a lot of output from several other commands
with lesser information value...

Change-Id: Id713e45e123c7f5aff36f6e55e081263efb35d9f