OSPD Undercloud fail when .rc != 0 otherwise skip
Usage of 'failed_when' magic is incorrect here,
it's not interchangeable with when at all!

It applies to (overrides) the result of the 'fail' task itself,
not to when the fail task should be executed.

It still overrides the 'failure' generated by fail task,
or pretty much it masks it always as 'success => [ok]' unless
rc of undercloud is nonzero.

But it causes that the 'fail' task and it's msg are always
printed, even in case of successful undercloud deploy,
which is more then confusing
(to see 'hey this failed :(' and run continuing).

There should be just regular when:
  - fail .. when deploy.rc != 0
which will:
- do red FAILure with msg  if deploy failed
- do lightblue SKIPped  if deploy succeeded

As this may be confusing (the failed_when != when together with fail
task), it's like:

> prev_task.rc = X
>
> try:
>     raise Failure(msg=...)
> catch Failure as exc:
>     if prev_task.rc != 0:
>         raise
>     else:
>         echo 'ok msg=%s' % exc

while we want here:

> prev_task.rc = X
>
> if prev_task.rc != 0:
>     raise Failure(msg=...)

Change-Id: I6e1f2073a21dd160f53ee42bbe0d7d433a20125c