[quickstart_fixes] Quickstart images are filling up the virthost
Currently, quickstart images are stored in ~/, means /root. If we
don't clean up these images, host will quickly loose storage space
and won't be able to run VM's anymore (all will be in paused state
when this occurs). Cleanup might not be reached in later steps, this
makes cleanup as soon as possible after download.

This patch is result of debugging host seal20 which suffers from this.

Change-Id: I3dead522a53ca711d550fc16705e6ba163e6759b