[quickstart_fixes] Quickstart images are filling up the virthost
Currently, quickstart images are stored in ~/, means /root. If we
don't clean up these images, host will quickly loose storage space
and won't be able to run VM's anymore (all will be in paused state
when this occurs).

This patch is result of debugging host seal20 which suffers
from this.

Change-Id: I7a5239064229e1d58e80183929cdb9a1bcb052c8