Create db dump instead of raw binary backup
Before this change we collected the mysql data files,
as it is. This form is not too useful, because it
requires the same version of the db to be able to restore,
and the restoring process also not too easy.

The text version of the DB frequently useful even
without doing a database restore.

So, we are switching the per DB consistent sql text
backup.
Now we will compress it via xz, but we might keep change
this strategy and just relay on the outher tar compression.

Change-Id: Ib6cac1528a42839faff9144af9c7dc75c0c2911e