Add sleep=5 argument to 'restart systemd-hostnamed' task on packstack deploy.
In case packstack deployment performed in 'all in one' scenario,
InfraRed has the same host in both groups - controller and compute.
Once it reach the 'restart systemd-hostnamed' task, it tries to restart the
service twice on the host at the same time. As the result, the play failed.
'sleep' argument allows to wait the defined amount of time between the stop and start command.

Change-Id: Ibebde60e5b2af212c014e097061f259cabef257c