virsh: download/virt-customize just ONCE!
currently provisioner/virsh,
base image preparation is happening in nested loop
over each topology.nodes and then for each nodes.disks entry

with typical 3nodes, 1disk each, it means three times
so instead of 3 minutes it takes 9, redoing the same
steps again and again

this solves this for the price of few extra set_fact calls
with accumulation and map/selectattr filters

majority of difficulty here is that there are no lists
of disk, but dictionaries, so it limits usage of simple
map&selectattr.

better seems to be to impletement custom filter for this
but that's another story, until that at least this.

Change-Id: I57951fc284177679261b520241e8e4fdeff235bf