Fix OSP10 SSL environment files
This change adjusts the OSP10 environment files needed for SSL enabled
overclouds. It provides a predefined IP address for the public VIP via
the PublicVirtualFixedIPs parameter. It also passes the
tls-endpoints-public-ip.yaml environment file which starting with OSP10
contains the EndpointMap. Previously the EndpointMap was part of the
enable-tls.yaml environment.

Change-Id: I0629e66fe5deb7340a994bab658a248c036b2b00