BUGFIX Packstack VNC console fix
During the packstack installation 'Configure hostname' task is
running as the pre-config packstack installation task.
Once the packstack install is running, it sets the hypervisor (compute)
name as defined it the inventory file and sets one of the VNC parameters
as hypervisor FQDN name. For example - "compute1.qe.lab.tlv.redhat.com"
As a result, the config unable to resolve the name and connection to the
VNC console of the instance failed.

We unable to move the 'Configure hostname' task after the packstack run,
because in that case, packstack add additional fake hypervisor to the
list.

The fix changes the FQDN name to the IP address of the compute node.

Change-Id: I973fe7e9942499079ab000f1b3bc13b24029d618