Fix for packstack 'all-in-one' deployment failure on 'remove all repo files' task
In case 'all-in-one' deployment running, task 'gather repo file names to remove'
gather repo file names on controller and compute hosts which is one single host.
When it passes to the 'remove all repo files' task, it tries to delete the repo files
twice on the same host. As a result, the play failed.

Change-Id: I51d7e01035f26f329d96322e2f339a8f3daa87c3