Introduce templating system to external network workflow
We need to have templating system in place to be able
to create external networks for various topologies.
Until now only virthost with hard-coded values was supported.

Now it works for BM and VH topologies, gathers data from TripleO
templates and it is idempotent (relatively).

Update: Added support for ipv6 too

Change-Id: I4a83ad3d0e01bf076bb088dbfc260ba7953b40ce