Revert "BUGFIX Set minimun for flavor disks and cpu"
This caused issue in cases where disk is actually 20G
(most common case) it needs to be 19 or so ... (the default value),
but attempts to lower this default min disk value
expose additional misbehaviour of this code.
See failing patchets/attempts in
https://review.gerrithub.io/#/c/294691/ (picks disk size 1 instead of 35
as disk size is 40 there: max(40-5, 1) != 1).

Until we understand what is going wrong in this code,
it has to be reverted to not cause additional failures.

This reverts commit e18ada9314db991e942443cdb24f40a7d9e02179.

Change-Id: I9aa79cb36480d25c4b0ec6699f31f618fecadea3