[Bugfix] Remove DNS code on UC
Gathering data from resolv.conf is unreliable and recently breaks
approx. 50% of CI, because suddenly dns6 entry occurs there
which is used as DNS server (10.0.0.1) for ctlplane network.

In some cases we have ipv6 value picked by that code, which breaks
neutron commands expecting ipv4 dns servers to be set.

Anyway, this code behaves weirdly, so far I'll remove it and see
if it is needed. If yes we have to come up with more reliable way.

Change-Id: Ibbe1c6229415895bee693a240047008de8cba380