More durable ssh connections
ssh timeout just helps when the connection is delayed,
but it does not helps, when any transient network issue
causes any kind of TCP/IP reject response from
any networking element or temporary from the target OS.

This change reduces the connection timeout to the default 10 sec,
but instead of one connection attempt, we will do 32.

It will increase the give up time, but we will not
loose a job, just because of temporary/transient anomaly.

Change-Id: Ia13269290b2db4d170a4ea469a09ee5cad62eb64