HOT FIX - remove password from "add_host"
When Ansible is using SSH, it looks first for the ansible_pass value
before looking for ansible_ssh_private_key_file value.
This patch removes the unnecessary set of password for the 'stack'
user since we already set the key for it.
This line causes Ansible to fail on machines that don't have the
'sshpass' package installed.

Change-Id: Ied028e5021c011081b98632a65d1fe34ff70e0a6