instance_ha: Make sure neutron-openvswitch-agent is started
neutron-openvswitch-agent is not managed by pcmk on RHOS10. But for Instance HA
the pcmk resource must be created for managing the service on computes which
results in the service being sometimes stopped by pcmk on controllers after
the resource is created and before proper location constraint (to manage the
service only on computes) is created.

Change-Id: I6dbafa38dadcf7a834702cede456243b3eb8ee51