Restructure virsh playbooks
This patch restructure the playbooks of the virsh provisioner.
The intent is to have a more "English like" understanding of the
stages of provisioning such that any person reading the playbook
can understand what it does.
This also introduce "injection points" when a user can "override"
any step with a custom step of his own.

Change-Id: I40d0766c180089aff37011dd0b5c13664f4537dc