HOTFIX - move var definition into block
For an unknown reason, the 'node_types' var does not get evaluated
properly causing our jobs to fail.
When trying the same job invoke locally, everything works as it should
This patch tries to move the vars definition inside to try give
a "quick fix" to this problem while we are trying to understand the
origin.
If this doesn't work, please revert

Change-Id: If164ab18b1d457bb9b7d45c4e58a63b178177228