Defines defaults for topology and adds docs
This patch moves all our topology files into a "default" folder such
that a user can create his own custom topology files while still
be able to use the defaults.
An example workflow is as follows:
First, we look in var/topology/nodes/<file> for the var file
if not exist, we default to the defaults/topology/nodes/<file>

Fixes __LISTYAML__ lookup to search defaults dir as well and build a set
of unified list of files from both locations.
Fixes unittest accordingly

Create tasks file for priority lookup and reuse it in virsh network and
nodes vars.

It also adds documentation about the usage of topology files
Change-Id: Ic4566802d17ea4305e4e1af277e40d3617b749ef