Remove 'when' statement from 'block'
This patch removes the behaviour when the DVR playbook was automatically
added as part of the Ansible pit-fall of using a 'when' on a 'block'.
A 'when' statement does not work on a 'block' and shouldn't be used.
This patch removes any 'when' statements on blocks into dedicated plays
and puts an 'include' statement to avoid running it.
This also fix the DVR code that was running since it had a 'when'
on the 'block' statement.

Change-Id: Icd2861b015263f91feeb018746826b80e7a7a0ec