don't merge - test - patching
This patch adds patch_rpm role to the installer part and runs the role
for packstack installation.

NOTE:
Plz, disregard the mess in args/8 for now. I need to make a cleanupp

TODO:
a) run the role for the ospd installation
b) allow patching a list of components (required for consolidated tempest nigtly jobs
that run on delta(puddle, latest_patches)

Change-Id: I74ad843e615c9dcf96a3227a56669678214bb1df