Reorg our tripleo-overcloud main playbook
This patch reorgs our main.yml playbook to align with tripleo-undercloud.
It adds the injection points to overcloud and moves tasks into logical
"steps".
This patch also removes the "selinux" task that changes the selinux
on the overcloud since we should not touch the deployment.
It also removes the "dump facts" which lost relevance when we split
the deployment to undercloud and overcloud.

Change-Id: Ie7bbca1733f910c7dc90c6bbf1cbc0a26a01fc7a