Add gather facts to the restore option
This patch changes the gather facts to "yes" (previously "no) as
running restore without first gathering the facts will fail on ansible
not being able to use the "package" module.
The "package" modules decides which package manager to use based
on the OS distro, for that, you require to gather the facts about
the machine in question.

Change-Id: I23bdfcf04f7b08838cd2afa8e33f9db043652a3e