VLAN needs empty string NeutronTunnelTypes
For vlan there should not be NeutronTunnelTypes specified,
but when we put such key in env file,
we need to give then some value, in this case empty string
seems to be used most commonly.

As this code here is templated string, and we are in live code
in condition it has to be encapsulated in outer string (doublequotes),
and as we are in yaml the whole thing has to be also in quotes as string
(and escape sequence does not works in single quotes either).

RHOSINFRA-754

Change-Id: I56d65252530759332e4b5e2023bd0e1c43d31e7a