Backport fixes doublequotes
VLAN needs empty string NeutronTunnelTypes

For vlan there should not be NeutronTunnelTypes specified,
but when we put such key in env file,
we need to give then some value, in this case empty string
seems to be used most commonly.

As this code here is templated string, and we are in live code
in condition it has to be encapsulated in outer string (doublequotes),
and as we are in yaml where we already used single quotes
it has to be escaped too.

RHOSINFRA-754

NeutronExternalNetworkBridge no doublequoted ''

Change-Id: Id060efd8a3dd191f7dec4a2fb853b35172c6564a