Update patch set 1
Patch Set 1: Code-Review-1

Is there a reason to restore it?
If I understand it correctly, in IR1 the file was required (and copied as ansible.cfg). Now it's not included, which could mean that it's simply not needed.

Patch-set: 1
Reviewer: Luigi Toscano <1001398@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1