Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

@Luigi, 
Actually, this is not needed for personal use.
For CI and scaling, you want ansible.cfg but I'm not clear on the defaults.
I don't have a problem keeping an example file in the repo as long as it won't mislead users to think it's mandatory.

Patch-set: 1
Label: Code-Review=-1