Update patch set 1
Patch Set 1:

No, that's the path which is provided by python-tempestconf package actually.

It was pretty much test for 'is python-tempestconf installed, and where exactly'. Assumption was that if tempest is installed, it will be installed too - if it's new version. But now deps got broken/reversed (as there are then some circular deps, until some other packages in chain get build or such, something about repo bootstrap building:

%if 0%{?repo_bootstrap} == 0
Requires:     python-tempestconf
%endif

Patch-set: 1