Update patch set 2
Patch Set 2:

At Red Hat, we are currently bound to support not only RHOS 6 but also RHOS 5 so adding a parameter that limits the range of version is a serious regression.

This would be what people call "planned obsolescence" and it doesn't help us at all. 

If IR would raise a warnings when param values are unknown (not officially supported) this would be feature, sadly it gives a fatal error.

In current state, this is a CRITICAL BUG as it does not only prevent using older versions but it also blocks us from using future versions without modifying the IR code.

On the funny side that like making a car that stops after the warranty period. Samsungs did something like this with its printer cartridges and it costed them $5M in compensations.

I don't know how the specs were implemented so what I did was to add version names that I am already aware that they will exist in the future. If you know how to make this param verification only a warning better, it would avoid future CR that could have been avoided.

Patch-set: 2