Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> My 2 cents on that:
> 
> * Let's add 6 as it's far from EOL
> * If the framework doesn't support 13 then why adding it? I do think it's a good idea to output something like "this version is not supported at the moment".

Indeed, a warning would be harmless but I don't know how to implement it. My current need is only to add 6 but considering the accelerated development model that we have, I would say that it may be better not prevent users from specifying new params.

Patch-set: 2
Label: Code-Review=0