Update patch set 3
Patch Set 3: Code-Review-2

Unfortunately, I have to decline this until I see a working run of packstack with version 6. The reason it is in EOL branch, is that it required several differences that were too much to support.
If you can prove that it works, we can merge this.

I'm marking this as -2 so it's not merged by accident without proof that it works. Feel free to send me a private proof and we can merge this

Patch-set: 3