Adds infrared_cwd ansible parameter passed when executing ansible
This is required by Octario plugin in order to expand relative path
to the full path on the system where component lives.

Without this change it's not possible to work with relative paths:

$ ir octario --t pep8 ../openstack/component

Above will try to find component in the openstack directory one
level up from octario plugin while user may be in very different
one.

Change-Id: I9d244a57cb83d8a54834b744e261f1f4e6703a99