Add NovaPlacement and MySQLCLient service profiles to composable roles
NovaPlacement is required in Ocata, without the service it's not
possible to even schedule an instance. It was not required in
Neutron but recommended so let's add it to controller role.

MySQLClient profile is introduced for Ocata and can be used for all
roles (especially for the one with haproxy/vips) but when ssl used
then with all roles (Added to galera/messaging roles since we support
these roles separated from Ocata).

Change-Id: I3d3e14c7b6dcfe0c2794e4481e1dbe031b4febfd