Update patch set 1
Patch Set 1:

Please reconsider the -2.
The mechanism here is general and it prevents future changes of the name of the package (which will happen). And as you can see, the versions are not aligned: the ironic-inspector plugin is available only from 11, so even if 7: and 8: are removed, the ironic_inspector suite would have a 10: "" anyway. And it will happen again when more plugins are supported.

Also, most of the suites were valid also for the previous versions, only the regexp was different. Before the plugin was available, data_processing had tempest.api.data_processing as regexp, now it's sahara_tempest_plugin.tests[.api]. This is not captured now because old versions were not too supported, but I'd prefer not to duplicate each suite for the old version when a simple regexp can capture both case, and the configuration for the plugin "" tells the system to not install it.

Patch-set: 1
Reviewer: Luigi Toscano <1001398@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0