Update patch set 2
Patch Set 2:

As mentioned on ticket:
User is expecting plugins (ironic for example) to be installed and executed. If Ironic plugin is unsupported for installed version, it should fail with proper error, and not skip and run no-tests.
On the other hand:
If you want to introduce version-specific rpm names, that's acceptable IMO.

Patch-set: 2