Update patch set 1
Patch Set 1:

The source tarball present in version >7 is quite old [https://github.com/redhat-openstack/tempest/tree/openstack-tempest-kilo-20151020] with respect to https://github.com/redhat-openstack/tempest/tree/kilo . and in the current 7 tempest rpm clean script is broken. I created a scratch build by exposing the tempest binaries but it not feasible.

 > I understand the version >7, but is there no way to use it when
 > deploying from git?
If we use kilo branch of redhat-openstack/tempest from git things works fine
for openstack-version-7 as i have tested it locally. or we need to create a new tag for kilo and update the source and spec in RPM or we can stick with git.

Patch-set: 1