Update patch set 17
Patch Set 17: Code-Review+2

I tested the patch and it worked fine for me. One thing that we need to do after the upgrade is to upload the new images provided by the updated rhosp-director-images package and run 'openstack baremetal configure boot'. It'd be great if we could do this in this patch but we can do it in a follow up patch as well.

Patch-set: 17