Use plain testr with rhos-7 tempest.rpm
The scripts what we had at those old ages
are very different to the ones what we are using today,
this change does not implements skipfile (blacklist file)
support for 7, but it is able to pass a smaller white list.

Unfortunately rhos-7 tempest-lib conflicts with ostestr,
at all sensible version, both package contains
the subunit-trace executable.

Change-Id: I12ad7ffd01b6fa127c7d2a05667829bdc594b4a2