Update patch set 3
Patch Set 3:

I don't like it, it's a too big chunk code just for 7. I think that there are possible solutions (an updated python-os-testr, even if it was a special package from copr too, or updating tempest-lib to a version without the conflict). Unfortunately this needs to be fixed, so I'm going to merge it nevertheless. If someone else want to propose a different solution, it should be introduced without regressions compared to the status after merging this patch.

Patch-set: 3