Place Controller role first in the roles_data.yaml
According to the [1], Controller role should be
placed at the start of the roles_data.yml. First role is
treated as primary role.

This patch adds the 'primary_role' flag. Role marked with that
flag will be listed first in the template.

Updated all the controller roles to be primary.

RHOSINFRA-951

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1436769

Change-Id: I2ade05cd5b0b3fbea05e3e550e9ee02dbc3c1dac