Add a fixed common path to be used as shared library
Add a directory to the search path of roles, filter and callback plugins
which points to a directory in the infrared code. This directory can contain
shared items of the said types for all the InfraRed plugins, reducing code
duplication.

Small drawback: the variables defined internally are then always defined.
This means that corresponding configuration keys inside ansible.cfg will
always be ignored.
Please note that  still possible to add additional paths for those items
using the environment variables, whose values are added and not overwritten.

Change-Id: I084819fdd38eca96b7cb26de4670e1ba2fc8950f