Add a fixed common path to be used as shared role library
Add a directory to the search path of roles which points to a
directory in the infrared code. This directory can contain shared
roles for all the plugins, reducing code duplication.

Small drawback: the variable ANSIBLE_ROLES_PATH is then always defined
and this means that roles_path inside ansible.cfg will always be ignored.
It is still possible to add additional role paths through
ANSIBLE_ROLES_PATH.

Change-Id: I084819fdd38eca96b7cb26de4670e1ba2fc8950f