Update patch set 1
Patch Set 1:

(4 comments)

pbr does solve the issue with recursive packaging, or I wouldn't have made the change to switch to pbr.

It seems we have no much space of negotiation. With the current direct usage of Ansible API, this seems to be the only way to implement this change. If you want to switch to calling ansible-playbook instead, that would simplify a lot of code.

Patch-set: 1