By default, deploy tempest from RPMs
We want to make sure packaged tests shipped with OSP actually pass, and
keeping git test env deployment as default is in conflict with that
goal. This patch makes the switch. Users that still want to use git can
explicitly define it.

This is also somewhat important right now since git deployment method is
not version aware, so it always deploys tempest tests from master
branches of projects, irrespective of overcloud release, which is doomed
to fail hard because some projects (like neutron) add tests that target
behaviour specific to a particular release.

Change-Id: Iee9d8fd5e7a85d23b25c182c549fa50dc4592e83