Update patch set 1
Patch Set 1:

Even if I totally agree with the last sentence and with this change (I've been pushing for it myself, even if on the definition of the jobs that we use), nothing prevents from adding a way to checkout specific branches for each plugin. (That said, most of the issues will be luckily solved when the spec about putting Tempest plugins in separate repositories will be accepted).

Patch-set: 1