Update patch set 7
Patch Set 7:

@Yurii Prokulevych

If we remove the part for image updade, update is the same as normal upgrade. The difference is that update have to install repos for the same version.

If I'm right, can we just add a flag to normal upgrade and when this flag is provided we will not install repos for next version but for the same one.

Update of the images can be done after Undercloud update with:

ir tripleo-undercloud -v --images-task rpm --images-update yes

Patch-set: 7