Update patch set 1
Patch Set 1:

What are you going to use in the help screen for each plugin then? All the "help" keys for all the subparsers associated with that plugin?

My point is that the value of description and help is currently duplicated only because each plugin currently in InfraRed has only one element under subparsers. Is it intended to be this way?

Patch-set: 1
Reviewer: Luigi Toscano <1001398@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0