Update patch set 5
Patch Set 5:

I'm not sure what could be the automagic configuration option. If you want to test the "base" python-tempestconf with some extra patch (which is the reason for this parameter), you could use tempestconf as deployed but then you need a parameter for the patch anyway.

Unless you pre-populate python-tempestconf where InfraRed expects it, so that you don't close it again, and assume that it's fine (i.e. don't clone it). It breaks if you change the tempest_conf directory, though; but you can force a new value from the command line, it's an internal variable. Not sure.

Patch-set: 5
Label: Code-Review=0