Update patch set 4
Patch Set 4:

Saying it here too: there is no reason to not set the name of the file at the beginning (as proposed in https://review.gerrithub.io/#/c/360142/) but using a timestamp (like 201705091423, using UTC time). You can get the last two results with ls (because the timestamp in almost-ISO format guarantees ordering).

Patch-set: 4