Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Saying it here too: there is no reason to not set the name of the file at the beginning (as proposed here) but using a timestamp (like 201705091423, using UTC time). You can get the last two results with ls (because the timestamp in almost-ISO format guarantees ordering).

Based on your approach it would be acceptable to append TIMESTAMP to any file generating during built. This will avoid overwrites at the expense on UX and spending a serious number of hours trying to implement file matching algorithms for diffing (because the result is almost never on the same file). 

I think that you forget that the TIMESTAMP was an workaround for not overwriting files, not a solution without side-effects, like changing the current filename. I see the TIMESTAMP as a step back from UX point view because it hits me with some additional complexity just because we want to avoid overwrites on 1% of the builds that have more one execution.

That's why I think a good implementation addresses the special case without affecting current behaviour.

To give you much easier to understand example: imagine that you do generate two or three pip-freeze reports during job execution. If you use timestamps you will not be able to easily compare how they differ between build #100 or #200. If you generate pip-freeze.txt, pip-freeze-2.txt, ... a simple diff 
will report you what changed. There is some beauty in keeping things simple. 

It would work mentioning that while debugging I hate having to deal with shell auto-complete on files with timestamped suffixed, it makes much easier to pick the wrong file.

Also, if I send someone a link to Jenkins tempest-2017....xml, they will have no idea if this was 1st, or 2nd execution. The URL will tell them nothing, they will have to navigate to the list of artefacts and look for which one of them is.

Patch-set: 1
Label: Code-Review=0