collect-logs: report selinux/oom/segfault for UC
Log collection, when counting selinux/oom/segfault
events, is explicitely excluding tester node,
as that was in past completely separated node,
which had nothing to do with product itself.

(And sometimes still could be the case,
 depending on test suite/tools used.)

There is already case of packstack all-in-one,
where tester is included, when it is also labeled controller.

But in case of OSP-D, undercloud is used as tester node too,
though it's issues should be reported,
and so this makes counting of issues to also apply to UC node.

Change-Id: I4d7ef57d0960276ae249361948ed5708bbdd2607
(cherry picked from commit 635a31969eea8775dc5189377091a22c9d1c3c23)