Update patch set 11
Patch Set 11: Code-Review-1

(2 comments)

Please make sure the user is able to specify arbitrary registry as configuration option instead of hardcoding one.
 --docker-registry <url>

Please make sure a version specifier link is not hardcoded, for example do not use 'latest_containers', we must be able to run with older versions.

Maybe adding an --extra-tempaltes <urls> option ?  Which might not be container specific feature.

Looks like you do not need to know on external url, to know which puddle was used on the undercloud, is it good for you ?

Patch-set: 11