Allow the vms to use the hosts random sources
Since el7 possible to provide better entropy
to the machines even without having a hardware rng.

Since el7.2 it is also possible to give non blocking, but
less random sources to the vms, Because, we do not
really care about the quality of these numbers in this
setup, we will use the urandom.

RHOSINFRA-1107

Change-Id: Ia43b243a98523d7d4b478773ff2b90e702ada24b