WIP Skip unused nodes after deploy
For some cases the number of overcloud nodes used in deployment and the
number of provisioned nodes can differ (for example, for scale up tests).
This patch adds 'unused' inventory group which holds all the provisioned
nodes which we not used for the overcloud deployment. So user can skip
those nodes in the post deployment tasks.

RHOSINFRA-1144

Change-Id: I2be34411bd4f422307bf9756900f02da1fee6628