Use smaller default pool in small envs
Ceph by default requires 3 node (osd)
for being able to create 3 replicas.
In smaller dev systems
We have only 1 or 2 node(s) in smaller dev systems,
updating the default accordingly.

The tripleo does not do it for you anymore,
it is kind of non-production setting,
so I will not file bug because of this.

Change-Id: I2b7fed0e49d118e4403c48c272f101af1a4619e3