Revert "Allow UC's subscription to CDN for minor update."
This patch broke the CDN nightly job.
The 'rhel-7-server-rhceph-2-osd-rpms' repos is not included in the 'Red Hat OpenStack Platform' subscription that in use in the job (as required in the documentation) and because of this the job fails when trying to enable this repo.
The repos that have been added in this patch should be enabled on the overcloud nodes only by the doc.

This reverts commit d1d9e5f51179c1912d90ff5b7f65b3ca77017957.

Change-Id: I66866e9c995fca1b9aedb5c779ab4967dfae009c