Fixes workspace cleanup failure when no '.active'
The workspace cleanup method calls the workspace delete method.
The workspace delete method used to delete the '.active' file from the
workspaces dir if the workspace that aimed to be deleted was active.
If the cleaned-up workspace was checked up before the cleanup and the
environment variable 'IR_WORKSPACE' was set, the activate method raise
an error and a new '.active' file wasn't created.
After that when try to delete an active workspace (defined in
IR_WORKSPACE), the delete operation failed becuase the '.active' file
was missing.
This patch re-activates only workspace that was active NOT by
'IR_WORKSPACE' env var.

RHOSINFRA-1191

Change-Id: I30e9e4d8317b7de674fd83266e7bb2dd1a371efe