Fixes workspace cleanup failure when no '.active'
The cleanup method has tried to activate the given workspace if it was
the active one and because as part of the cleanup process the
'workspace delete' was called, the '.active' file was deleted.
When the environment varialble 'IR_WORKSPACE' was set, the workspace
checkout call was failed anf because of that no new '.active' filw was
created.
After that, any workspace cleanup/delete operation was failed becuase
the '.active' file was missing.

Changes in this patch:
----------------------
1. The '.active' file will never be deleted when calling to workspace
   cleanup.
2. The '.active' file will be deleted only if it contains the same name
   as provided to the workspace delete method.
3. Add a check that the '.active' file exists before deleting it.

RHOSINFRA-1191

Change-Id: I30e9e4d8317b7de674fd83266e7bb2dd1a371efe