Update patch set 1
Patch Set 1: Verified+1

Arie, I don't see what else would be necessary, my comment is as clear as it can be "octario plugin".

Btw: "octario ir plugin" is incorrect in English because is gives impression that IR is a plugin of octario, which is not true. 

Order is important in this case, so the only valid descriptions are: "infrared octario plugin" or short version "octario plugin"

Patch-set: 1
Label: Verified=+1