Fix disks list for RHOSP 11+
To collect information about storage node disks we currently read
the instackenv.json file for the disks property.

Currently ir goes through the list of nodes and takes the first node
which can be not a storage nodes. This patche fixes that issue by
querying only the first storage node for the disks property.

Also renamed the file to fix a typo.

Change-Id: I63a797b62eded659ff10f6cf6f8d2b163ef9ca40