Do not run overcloud upgrade steps starting with OSP12
Starting with OSP12 we're planning to use tripleo-upgrade repo,
provided as a separate Infrared plugin, for upgrading the under/
overcloud. This change allows us to keep using the tripleo-overcloud
plugin for setting up the repositories on the overcloud nodes.
Currently it fails because there is no tripleo-overcloud/tasks/upgrade/12/main.yml
file existing.

Change-Id: I42350153ab72364cc7d28f271c258c3c09224c5f