Update patch set 2
Patch Set 2:

(1 comment)

I'm not familiar enough with yaml/templating/linting for how this change interacts for where it changes.  I can't give a +1 for that bit, but +1 by me for enabling yamllint for gates.  Not sure if there is significant benifet to including a Makefile in here though.

Patch-set: 2