Adds 'yedit' Ansible module to tripleo-overcloud
Added the [1] 'yedit' Ansible module to tripleo-overcloud plugin.
Replaced the use in 'lineinfile' with 'yedit' where editing the network
deployment file in this plugin.
The need in this module raised when a user has supplied his own network
deployment (YAML) file with an indentation that different from the one
was in 'lineinfile' value.

[1] https://github.com/kwoodson/yedit

RHOSINFRA-1231

Change-Id: Ia41777096d177850f9c909e48f70db59f31ca9b4