Update patch set 3
Patch Set 3:

When is run under Jenkins it will create descriptions like:
"ir: https://rhos-qe-jenkins.rhev-ci-vms.eng.rdu2.redhat.com/job/.../2/ at <date>"

When you run it outside Jenkins it will produce descriptions like:
"ir: ssbarnea@ssbarnea-macbook at <date>"

The explanation for difference is that under CI we need/want to know which job allocated the resource, user and host are less important because we can find them on the job console anyway.

Still, when run manually we do not have a job so we put username and hostname, two things that should be enough to give an idea about who created that resource.

Date is useful for measuring "age".

Here is the behaviour before this patch http://s3.sbarnea.com/ss/Access__Security_-_OpenStack_Dashboard_2017-07-21_13-28-51.png -- which is not of much help.

Patch-set: 3